+++ examples/resources/catalystcenter_update_authentication_profile/import.sh
@@ -1,0 +1,1 @@
+terraform import catalystcenter_update_authentication_profile.example "<authentication_profile_id>,<authentication_profile_name>"