+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.7 (unreleased)
 
+- Add import and data source to `catalystcenter_update_authentication_profile` resource
 - Remove default value from `interface_name` in the `catalystcenter_wireless_profile` resource 
 - Add import to `catalystcenter_assign_templates_to_tag` resource
 - Add `catalystcenter_assign_devices_to_tag` resource and data source